scripts: west_commands: create_board: nrf54l: Update files
Adds missing nodes and removes duplicate settings

+++ scripts/west_commands/create_board/ncs_create_board.py
@@ -158,7 +158,8 @@ class NcsCreateBoard(WestCommand):
 
         # board-level files
         tmpl = TEMPLATE_DIR / series / "pre_dt_board.cmake"
-        shutil.copy(tmpl, out_dir)
+        if tmpl.exists():
+            shutil.copy(tmpl, out_dir)
 
         tmpl = TEMPLATE_DIR / series / "board-pinctrl.dtsi"
         shutil.copy(tmpl, out_dir / f"{ input['board'] }-pinctrl.dtsi")
